pulled to Windows and Visual Studio, removed 2nd day07 attempt as fixed first

+++ src/day03/day03.cpp
@@ -139,7 +139,7 @@ namespace Puzzle2
 
         while (inf)
         {
-            for (uint i{ 0 }; i < 3; ++i)
+            for (size_t i{ 0 }; i < 3; ++i)
             {
                 std::getline(inf, elfGroup[i]);
